Update GitHub Actions workflow to optimize Android build settings for Safarnak app
This commit modifies the build-and-deploy.yml file to enhance the Android build process by enabling the Gradle daemon for improved performance and adjusting the API level and build tools to the latest versions. Additionally, it streamlines the dependency installation process by using the --prefer-offline flag. These changes aim to improve build efficiency and reduce build times in the CI/CD pipeline.