Update GitHub Actions workflow to improve Android build configuration for Safarnak app
This commit modifies the build-and-deploy.yml file to enhance the Android build process by accepting Android SDK licenses and adjusting the installation of essential components for APK building. These changes aim to streamline the build process and ensure compliance with SDK requirements, ultimately improving the efficiency of the CI/CD pipeline.