Update GitHub Actions workflow to enhance permissions and cache settings for Safarnak app
This commit modifies the build-and-deploy.yml file to add write permissions for contents and packages, ensuring smoother operations during the CI/CD process. Additionally, it changes the Gradle job summary setting to 'never' for improved clarity in build logs. These updates aim to optimize the build process and enhance the overall efficiency of the CI/CD pipeline.