Update GitHub Actions workflow to enhance caching and token management for Safarnak app
This commit modifies the build-and-deploy.yml file to set the Gradle cache to read-only, improving cache integrity during builds. Additionally, it updates the GitHub token handling for release creation and asset uploads, ensuring that the appropriate token is used for enhanced security and permissions. These changes aim to optimize the CI/CD pipeline and improve the reliability of the build process.